+++ server/collaboration/AuthenticationExtension.ts
