+++ server/collaboration/PersistenceExtension.ts
